requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
tensorflow==2.13.0
keras==2.13.1
hmmlearn==0.3.0
ccxt==3.1.54
requests==2.31.0
python-dotenv==1.0.0
jupyter==1.0.0
plotly==5.16.1
joblib==1.3.2
tqdm==4.66.1
pymongo==4.5.0
ta==0.10.2
statsmodels==0.14.0
transformers==4.32.1
torch==2.0.1
cybotrade-datasource==0.1.0
dash==2.13.0
dash-bootstrap-components==1.5.0
kaleido==0.2.1
nbformat==5.9.2 